Separate ColorPicker and ThemeParameters sliders
