+++ packages/fluent/src/components/ColorPicker.types.ts
@@ -2,26 +2,10 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
-import type {
-	IColorPickerProps,
-	IColorPickerStyles,
-	ISliderStyles,
-} from '@fluentui/react'
-import type { Theme } from '@thematic/core'
-import type { CSSProperties } from 'react'
+import type { IColorPickerProps } from '@fluentui/react'
+import type { SchemeParams } from '@thematic/color'
 
-export enum ColorPickerLayout {
-	PickerOnly = 'pickeronly',
-	SideBySide = 'sidebyside',
-}
-
-export interface ColorPickerStyles extends IColorPickerStyles {
-	container?: CSSProperties
-	slider?: ISliderStyles
-}
-
-export interface ColorPickerProps extends Omit<IColorPickerProps, 'onChange'> {
-	onChange?: (theme: Theme) => void
-	layout?: ColorPickerLayout
-	styles?: ColorPickerStyles
+export interface ColorPickerProps
+	extends Omit<IColorPickerProps, 'color' | 'onChange'> {
+	onChange?: (params: SchemeParams) => void
 }
