+++ packages/fluent/src/components/ThemeParameters.types.ts
@@ -2,15 +2,16 @@
  * Copyright (c) Microsoft. All rights reserved.
  * Licensed under the MIT license. See LICENSE file in the project.
  */
+import type { ISliderStyles } from '@fluentui/react'
 import type { SchemeParams } from '@thematic/color'
 import type { CSSProperties } from 'react'
 
-export interface ColorPickerButtonStyles {
-	label?: CSSProperties
+export interface ThemeParametersStyles {
+	container?: CSSProperties
+	slider?: ISliderStyles
 }
 
-export interface ColorPickerButtonProps {
+export interface ThemeParametersProps {
 	onChange?: (params: SchemeParams) => void
-	label?: string
-	styles?: ColorPickerButtonStyles
+	styles?: ThemeParametersStyles
 }
