+++ packages/fluent/src/components/index.ts
@@ -10,3 +10,5 @@ export * from './ScaleDropdown/ScaleDropdown.js'
 export * from './ScaleDropdown/ScaleDropdown.types.js'
 export * from './ScaleTypeChoiceGroup.js'
 export * from './ScaleTypeChoiceGroup.types.js'
+export * from './ThemeParameters.js'
+export * from './ThemeParameters.types.js'
