+++ packages/webapp/src/pages/MainPage/components/ThemeEditor/panes/ColorPickerPane/ColorSelection.tsx
@@ -3,22 +3,21 @@
  * Licensed under the MIT license. See LICENSE file in the project.
  */
 
-import type { Theme } from '@thematic/core'
-import { ColorPicker, ColorPickerLayout } from '@thematic/fluent'
+import type { SchemeParams } from '@thematic/color'
+import { ColorPicker, ThemeParameters } from '@thematic/fluent'
 import type { FC } from 'react'
 
 export interface ColorSelectionProps {
-	onThemeLoaded: (theme: Theme) => void
+	onParamsChanged: (params: SchemeParams) => void
 }
 
-export const ColorSelection: FC<ColorSelectionProps> = ({ onThemeLoaded }) => {
+export const ColorSelection: FC<ColorSelectionProps> = ({
+	onParamsChanged,
+}) => {
 	return (
 		<div className="color-controls-area">
-			<ColorPicker
-				onChange={onThemeLoaded}
-				layout={ColorPickerLayout.SideBySide}
-				styles={{ container: { width: 400 } }}
-			/>
+			<ColorPicker onChange={onParamsChanged} />
+			<ThemeParameters onChange={onParamsChanged} />
 		</div>
 	)
 }
